restructured the folders
