+++ src/app/pitwall/dashboard/[id]/dashboards/dashboard-standings.tsx
@@ -5,37 +5,17 @@ import { SingleClassDetails } from "@/components/car-class-details/single-class"
 import { Standings } from "@/components/standings/standings";
 import { TrackMap } from "@/components/trackmap/trackmap";
 import { selectCurrentTrackSession, useSelector } from "@/lib/redux";
-import { useEffect, useState } from "react";
-import ReactGridLayout, { Layout, WidthProvider } from "react-grid-layout";
-import { DashboardCard } from "./dashboard-card";
-
-const ResponsiveGridLayout = WidthProvider(ReactGridLayout);
+import { useState } from "react";
+import { Layout } from "react-grid-layout";
+import Dashboard from "../components/dashboard";
+import { DashboardCard } from "../components/dashboard-card";
+import { DashboardComponent } from "../components/dashboard-component";
 
 const STORAGE_KEY: string = "dashboard_standings_layout";
-
-class DashboardComponent {
-  public id: string;
-  public content: React.ReactNode;
-  public title: string;
-  public description: string;
-
-  constructor(
-    id: string,
-    title: string,
-    description: string,
-    content: React.ReactNode,
-  ) {
-    this.id = id;
-    this.title = title;
-    this.description = description;
-    this.content = content;
-  }
-}
 
 export default function DashboardStandings() {
   const session = useSelector(selectCurrentTrackSession);
   const [carClassTitle, setCarClassTitle] = useState("");
-  const [layout, setLayout] = useState<Layout[]>();
 
   const defaultLayout: Layout[] = [
     { i: "carClassDetails", x: 0, y: 2, w: 4, h: 2, minH: 1 },
@@ -80,47 +60,11 @@ export default function DashboardStandings() {
     ),
   ];
 
-  useEffect(() => {
-    let savedLayoutJson = localStorage.getItem(STORAGE_KEY);
-    if (savedLayoutJson != null) {
-      let savedLayout: Layout[] = JSON.parse(savedLayoutJson);
-      setLayout(savedLayout);
-    } else {
-      setLayout(defaultLayout);
-    }
-  }, []);
-
-  function renderComponents() {
-    return layout?.map((layoutItem) => (
-      <div
-        key={layoutItem.i}
-        data-grid={layoutItem}
-        className="overflow-hidden"
-      >
-        {components.find((c) => c.id === layoutItem.i)?.content}
-      </div>
-    ));
-  }
-
-  const onLayoutChange = (layout: Layout[]) => {
-    localStorage.setItem(STORAGE_KEY, JSON.stringify(layout));
-  };
-
   return (
-    <>
-      {console.log("render", layout)}
-      <main className="text-sm">
-        <ResponsiveGridLayout
-          className="layout"
-          isDraggable={true}
-          isResizable={true}
-          draggableHandle=".drag-handle"
-          onLayoutChange={onLayoutChange}
-          resizeHandles={["se"]}
-        >
-          {renderComponents()}
-        </ResponsiveGridLayout>
-      </main>
-    </>
+    <Dashboard
+      storageKey={STORAGE_KEY}
+      defaultLayout={defaultLayout}
+      components={components}
+    />
   );
 }
